With the new implementation performance do not suffer.

bert-base-multilingual-cased
Old
Sync:  OV: **2123.490 FPS**
Async: OV: 4719.860 FPS
New
Sync:  OV: **2208.642 FPS**
Async: OV: 4929.852 FPS

cointegrated_rubert-tiny2
Old
Sync:  OV: **2131.997 FPS**
Async: OV: 5586.107 FPS
New
Sync:  OV: **2079.225 FPS**
Async: OV: 5555.836 FPS

distilbert-base-uncased-finetuned-sst-2-english
Old
Sync:  OV: **1930.645 FPS**
Async: OV: 2890.250 FPS
New
Sync:  OV: **1894.172 FPS**


Async: OV: 5085.155 FPS

google/mobilebert-uncased
Old
Sync:  OV: **1825.989 FPS**
Async: OV: 4877.011 FPS
New
Sync:  OV: **1900.643 FPS**
Async: OV: 3900.510 FPS

ProsusAI/finbert
Old
Sync:  OV: **1864.090 FPS**
Async: OV: 3624.136 FPS
New
Sync:  OV: **1883.059 FPS**
Async: OV: 4039.700 FPS

rasa/LaBSE
Old
Sync:  OV: **2038.847 FPS**
Async: OV: 4391.581 FPS
New
Sync:  OV: **2158.708 FPS**
Async: OV: 4458.595 FPS

sentence-transformers/all-MiniLM-L6-v2
Old
Sync:  OV: **1863.720 FPS**
Async: OV: 5027.566 FPS
New
Sync:  OV: **1718.141 FPS**
Async: OV: 4857.865 FPS